backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.23.0
confection==1.3.3
cryptography==47.0.0
cymem==2.0.13
distro==1.9.0
emoji==2.15.0
fastapi==0.136.1
filelock==3.29.0
filetype==1.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
installer==0.7.0
Jinja2==3.1.6
jiter==0.14.0
langdetect==1.0.9
llvmlite==0.47.0
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numba==0.65.1
numpy==2.4.4
olefile==0.47
openai==2.33.0
packaging==26.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
preshed==3.0.13
psutil==7.2.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pypdf==6.10.2
pypdfium2==5.7.1
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.4.20
python-magic==0.4.27
python-multipart==0.0.27
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.5
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
thinc==8.3.13
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured-client==0.43.2
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
webencodings==0.5.1
websockets==16.0
wrapt==2.1.2
